variable name changes and blocking update
changed some variable names to ones that make more sense.
changed interaction between player and AI when doing hit checks.
added a collider and interface to the weapon to allow blocking from same direction when turned around, but this is probably not finished. Right and left blocking can still block up attacks. Not sure if we want that.